MANDELBULB : cleanup code -- animate
